requirements.txt
pdfplumber
psycopg2-binary
pandas
openpyxl
